implemented ControlPanel component in TypeScript and updated App.tsx to use it.
