+++ skyhigh_studios_assessment_toolbar_source_code_official/src/App.tsx
@@ -1,23 +1,14 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
+import ControlPanel from './components/ControlPanel';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
+        <h1>Control Panel Demo</h1>
+        <ControlPanel />
       </header>
     </div>
   );
